server/requirements.txt
alembic==1.13.1
authlib==1.3.0
fastapi==0.109.0
fastapi-sqlalchemy==0.2.1
llama-index==0.10.19
psycopg2-binary==2.9.9
pyjwt==2.8.0
pymilvus==2.3.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
sqlalchemy==2.0.28
tokenizers==0.15.0
torch==2.1.2
transformers==4.38.2
urllib3==1.26.18
uvicorn==0.26.0
itsdangerous==2.0.1
slack-sdk==3.27.1
llama-index-llms-ollama==0.1.2
llama-index-llms-openai==0.1.10
llama-index-core==0.10.20
llama-index-agent-openai==0.1.5
llama-index-readers-jira==0.1.3
llama-index-readers-slack==0.1.3
llama-index-vector-stores-milvus==0.1.6
llama-index-readers-file==0.1.11
llama-index-embeddings-huggingface==0.1.4


